requirements.txt
Django==5.0.6
pillow==10.3.0
djangorestframework==3.15.1
